% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_exclusions}
\alias{fetch_exclusions}
\title{Get item exclusions}
\usage{
fetch_exclusions(version = "current")
}
\arguments{
\item{version}{Redivis version tag of the \code{levante_metadata_items} dataset}
}
\description{
The item exclusion table, synced from Airtable. See
\code{\link[=apply_exclusions]{apply_exclusions()}} for its columns.
}
\examples{
\dontrun{
exclusions <- fetch_exclusions()
}
}
