% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-metadata.R
\name{fetch_item_mapping_id}
\alias{fetch_item_mapping_id}
\title{Get item UIDs mapped by item_id}
\usage{
fetch_item_mapping_id(version = "current")
}
\arguments{
\item{version}{Redivis version tag of the \code{levante_metadata_items} dataset}
}
\description{
Get item UIDs mapped by item_id
}
\examples{
\dontrun{
mapping_id <- fetch_item_mapping_id()
}
}
