% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-fit.R
\name{fit_task_models_multigroup}
\alias{fit_task_models_multigroup}
\title{for a given task and group variable, fit and record set of multigroup models}
\usage{
fit_task_models_multigroup(
  task_data,
  models,
  priors,
  task,
  group,
  registry_dir
)
}
\arguments{
\item{task_data}{dataframe with columns item_task, the value of subset_var,
and data, where data is trial-level data}

\item{models}{dataframe of models to fit, specified with columns nfact
(number of factors), itemtype (Rasch/2PL), invariance
(configural/metric/scalar)}

\item{priors}{list of priors, where names are parameter names (e.g. d, a1)
and values are vectors of length 3 (priorType, val1, val2)}

\item{task}{one value in item_task}

\item{group}{variable to use as groups (string or bare variable)}

\item{registry_dir}{string indicating directory in which to save models}
}
\description{
for a given task and group variable, fit and record set of multigroup models
}
