% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-fit.R
\name{fit_task_models_pooled}
\alias{fit_task_models_pooled}
\title{for a given task + subsetting variable/value, fit set of pooled models}
\usage{
fit_task_models_pooled(
  task_data,
  models,
  priors,
  task,
  subset_var,
  subset_val,
  registry_dir
)
}
\arguments{
\item{task_data}{dataframe with columns item_task, the value of subset_var,
and data, where data is trial-level data}

\item{models}{dataframe of models to fit, specified with columns nfact
(number of factors), itemtype (Rasch/2PL), invariance
(configural/metric/scalar)}

\item{priors}{list of priors, where names are parameter names (e.g. d, a1)
and values are vectors of length 3 (priorType, val1, val2)}

\item{task}{one value in item_task}

\item{subset_var}{variable to subset by (string or bare variable)}

\item{subset_val}{value to subset to (string)}

\item{registry_dir}{string indicating directory in which to save models}
}
\description{
for a given task + subsetting variable/value, fit set of pooled models
}
