% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt-fit.R
\docType{data}
\name{invariances}
\alias{invariances}
\title{mapping from invariance names to corresponding mirt argument}
\format{
An object of class \code{list} of length 3.
}
\usage{
invariances
}
\description{
mapping from invariance names to corresponding mirt argument
}
\keyword{datasets}
